+++ user_system.py
@@ -1,25 +1,20 @@
 from replit import db
+from extra import *
 
 class User():
 
         def __init__(self,username,password,history):
                 self._username = username
-                self._password = User.__encrypt(username,password)
+                self._password = encrypt(username,password)
                 self._history = history
 
-        def __encrypt(username,password):
-                toRet = ""
-                for i in range(len(password)):
-                        toRet += chr((ord(password[i])+1)^(ord(username[i%len(username)])))
-                return toRet
-
         def get_user(username,password):
-                if username in db and db[username][0] == User.__encrypt(username,password):
+                if username in db and db[username][0] == encrypt(username,password):
                         return User(username,db[username][0],db[username][1])
 
         def register(username,password):
                 if not username in db:
-                        db[username] = [User.__encrypt(username,password),[]]
+                        db[username] = [encrypt(username,password),[]]
 
         def log_scrape(self,event):
                 self._history.append(event)
